Extracted validateJsonField
